Move group containers to their own directory.
