+++ scripts/containers/group/GroupCreatePage.js
@@ -1,8 +1,8 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 
-import GroupCreate from "../components/group/GroupCreate";
-import * as BucketActions from "../actions/bucket";
+import GroupCreate from "../../components/group/GroupCreate";
+import * as BucketActions from "../../actions/bucket";
 
 
 function mapStateToProps(state) {
