+++ scripts/containers/group/GroupHistoryPage.js
@@ -2,9 +2,9 @@ import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { push as updatePath } from "react-router-redux";
 
-import GroupHistory from "../components/group/GroupHistory";
-import * as GroupActions from "../actions/group";
-import * as NotificationsActions from "../actions/notifications";
+import GroupHistory from "../../components/group/GroupHistory";
+import * as GroupActions from "../../actions/group";
+import * as NotificationsActions from "../../actions/notifications";
 
 
 function mapStateToProps(state) {
