+++ scripts/routes.js
@@ -15,9 +15,9 @@ import BucketEditPage from "./containers/bucket/BucketEditPage";
 import BucketCollectionsPage from "./containers/bucket/BucketCollectionsPage";
 import BucketGroupsPage from "./containers/bucket/BucketGroupsPage";
 import BucketHistoryPage from "./containers/bucket/BucketHistoryPage";
-import GroupCreatePage from "./containers/GroupCreatePage";
-import GroupEditPage from "./containers/GroupEditPage";
-import GroupHistoryPage from "./containers/GroupHistoryPage";
+import GroupCreatePage from "./containers/group/GroupCreatePage";
+import GroupEditPage from "./containers/group/GroupEditPage";
+import GroupHistoryPage from "./containers/group/GroupHistoryPage";
 import CollectionListPage from "./containers/collection/CollectionListPage";
 import CollectionHistoryPage from "./containers/collection/CollectionHistoryPage";
 import CollectionCreatePage from "./containers/collection/CollectionCreatePage";
